% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmgp.R
\name{bcmgp}
\alias{bcmgp}
\title{Bayesian Committee Machine (GP)}
\usage{
bcmgp(
  X,
  y,
  M = NULL,
  max_size = NULL,
  partition = "cluster",
  expert_weight = "varying",
  isotropic = NULL
)
}
\arguments{
\item{X}{A dataframe or matrix of predictors scaled to be between 0 and 1}

\item{y}{a reponse vector of length n}

\item{M}{Number of data partitions (defaults to \code{floor(sqrt(n))}).}

\item{max_size}{The maximum partition size. Ignored when \code{partition="random"}.}

\item{partition}{Data partition type; Either "random" or "cluster"; the latter uses a fast PAM algorithm to form the data partitions.}

\item{expert_weight}{How should the experts be weighted? Setting to \code{"uniform"} gives beta_i = 1 (as in Tresp 2000) while the default \code{"varying"} sets beta_i as formulated in Deisenroth and Ng 2015.}

\item{isotropic}{Logical. Should each input have the same lengthscale? (Defaults to \code{FALSE} when \code{ncol(X) <= 5} and \code{TRUE} otherwise).}
}
\description{
Many subset of data GPs are aggregated via the Robust Bayesian Committee algorithm of
}
\details{
Algorithm has complexity O(n(n/M)^2).
}
\examples{
X <- lhs::maximinLHS(100, 2)
f <- function(x) 10.391*((x[1]-0.4)*(x[2]-0.6) + 0.36)
y <- apply(X, 1, f) + stats::rnorm(100, 0, 0.1)
fit <- bcmgp(X, y)
}
\references{
Tresp, Volker. "A Bayesian committee machine." Neural computation 12.11 (2000): 2719-2741.

Deisenroth, Marc, and Jun Wei Ng. "Distributed gaussian processes." International conference on machine learning. PMLR, 2015.
}
