% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svecchgp.R
\name{fit_vcf}
\alias{fit_vcf}
\title{Plot Method for class mpgp}
\usage{
fit_vcf(
  fit,
  m.pred = 140,
  n.test = min(1000, round(nrow(fit$locs)/5)),
  scale = "parms",
  scorefun = ls
)
}
\arguments{
\item{x}{An object returned by the \code{mpgp} function.}

\item{...}{additional arguments passed to \code{plot}}
}
\description{
See \code{mpgp} for details.
}
\details{
Plot function for mpgp.
}
\examples{
X <- lhs::maximinLHS(100, 2)
f <- function(x) 10.391*((x[1]-0.4)*(x[2]-0.6) + 0.36)
y <- apply(X, 1, f) + stats::rnorm(100, 0, 0.1)
fit <- mpgp(X, y)
plot(fit)
}
