% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmgp.R
\name{plot.bcmgp}
\alias{plot.bcmgp}
\title{Plot Method for class bcmgp}
\usage{
\method{plot}{bcmgp}(x, ...)
}
\arguments{
\item{x}{An object returned by the \code{bcmgp} function.}

\item{...}{additional arguments passed to \code{plot}}
}
\description{
See \code{bcmgp} for details.
}
\details{
Plot function for bcmgp.
}
\examples{
X <- lhs::maximinLHS(100, 2)
f <- function(x) 10.391*((x[1]-0.4)*(x[2]-0.6) + 0.36)
y <- apply(X, 1, f) + stats::rnorm(100, 0, 0.1)
fit <- bcmgp(X, y)
plot(fit)
}
