% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpgp.R
\name{plot.mpgp}
\alias{plot.mpgp}
\title{Plot Method for class mpgp}
\usage{
\method{plot}{mpgp}(x, ...)
}
\arguments{
\item{x}{An object returned by the \code{mpgp} function.}

\item{...}{additional arguments passed to \code{plot}}
}
\description{
See \code{mpgp} for details.
}
\details{
Plot function for mpgp.
}
\examples{
X <- lhs::maximinLHS(100, 2)
f <- function(x) 10.391*((x[1]-0.4)*(x[2]-0.6) + 0.36)
y <- apply(X, 1, f) + stats::rnorm(100, 0, 0.1)
fit <- mpgp(X, y)
plot(fit)
}
