% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvm.R
\name{plot.rvm}
\alias{plot.rvm}
\title{Plot Diagnostics for RVM Objects}
\usage{
\method{plot}{rvm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rvm} as returned by \code{rvm()}.}

\item{...}{Additional plotting arguments.}
}
\description{
Plots diagnostic summaries for a fitted \code{rvm} model.
}
