% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svecchgp.R
\name{plot.svecgp}
\alias{plot.svecgp}
\title{Plot Method for class svecgp}
\usage{
\method{plot}{svecgp}(x, ...)
}
\arguments{
\item{x}{An object returned by the \code{svecgp} function.}

\item{...}{additional arguments passed to \code{plot}}
}
\description{
See \code{svecgp} for details.
}
\details{
Plot function for mpgp.
}
\examples{
X <- lhs::maximinLHS(100, 2)
f <- function(x) 10.391*((x[1]-0.4)*(x[2]-0.6) + 0.36)
y <- apply(X, 1, f)
fit <- svecgp(X, y)
plot(fit)
}
