% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcmgp.R
\name{predict.bcmgp}
\alias{predict.bcmgp}
\title{Predict Method for class bcmgp}
\usage{
\method{predict}{bcmgp}(object, newdata = NULL, samples = 1000, nugget = TRUE, ...)
}
\arguments{
\item{object}{An object returned by the \code{bcmgp} function.}

\item{newdata}{A dataframe of the same dimension as the training data.}

\item{samples}{How many posterior samples should be taken at each test point?
If 0 or FALSE, then the posterior mean is returned as a 1 x n matrix.}

\item{nugget}{Logical. Should the predictive variance include the nugget term?
Defaults to \code{TRUE}.}

\item{...}{Additional arguments to predict.}
}
\description{
See \code{bcmgp()} for details.
}
\details{
Predict function for bcmgp.
}
\examples{
X <- lhs::maximinLHS(100, 2)
f <- function(x) 10.391*((x[1]-0.4)*(x[2]-0.6) + 0.36)
y <- apply(X, 1, f)
fit <- bcmgp(X, y)
predict(fit)

}
