% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpgp.R
\name{predict.mpgp}
\alias{predict.mpgp}
\title{Predict method for class mpgp}
\usage{
\method{predict}{mpgp}(object, newdata = NULL, samples = 1000, nugget = TRUE, ...)
}
\arguments{
\item{object}{An object returned by \code{mpgp()}.}

\item{newdata}{A matrix/dataframe of predictors.}

\item{samples}{Number of predictive draws to return. If 0 or FALSE, returns
a 1 x n matrix of predictive means.}

\item{nugget}{Logical. If TRUE, include nugget / observation noise in the
predictive variance when available.}

\item{...}{Ignored.}
}
\description{
Predict method for class mpgp
}
