% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpgp.R
\name{predict.mpgp}
\alias{predict.mpgp}
\title{Predict Method for class mpgp}
\usage{
\method{predict}{mpgp}(object, newdata = NULL, samples = 1000)
}
\arguments{
\item{object}{An object returned by the \code{mpgp} function.}

\item{newdata}{A dataframe of the same dimension as the training data.}

\item{samples}{How many posterior samples should be taken at each test point? If 0 or FALSE, then the MAP estimate is returned.}

\item{...}{Additional arguments to predict}
}
\description{
See \code{mpgp()} for details.
}
\details{
Predict function for mpgp. Just a wrapper for predict.GP but matches the output expected by \code{duqling} package.
}
\examples{
X <- lhs::maximinLHS(100, 2)
f <- function(x) 10.391*((x[1]-0.4)*(x[2]-0.6) + 0.36)
y <- apply(X, 1, f)
fit <- mpgp(X, y)
predict(fit)

}
